+++ lib/types/interfaces.ts
@@ -1,71 +1,5 @@
-export interface Driver {
-    driverId: string;
-    name: string;
-    surname: string;
-    nationality: string;
-    birthday: string;
-    number: number;
-    shortName: string;
-    url: string;
-}
-
-export interface CurrentDriver extends Driver {
-    teamId: string;
-}
-
-export interface CurrentDriversF1apiResponse {
-  api: string;
-  url: string;
-  limit: number;
-  offset: number;
-  total: number;
-  season: number;
-  championshipId: string;
-  drivers: CurrentDriver[];
-}
-
-export interface Team {
-    teamId: string;
-    teamName: string;
-    teamNationality: string;
-    firstAppeareance: number;
-    constructorsChampionships: number;
-    driverChampionships: number;
-    url: string;
-}
-
-export interface DriverChampionshipApiResponse {
-    api: string;
-    url: string;
-    limit: number;
-    offset: number;
-    total: number;
-    season: number;
-    championshipId: string;
-    drivers_championship: DriverChampionshipEntry[];
-}
-
-export interface DriverChampionshipEntry {
-    classificationId: number;
-    driverId: string;
-    teamId: string;
-    points: number;
-    position: number;
-    wins: number;
-    driver: Omit<Driver, 'driverId'>;
-    team: DriverChampionshipTeamResponse;
-}
-
-export interface DriverChampionshipTeamResponse{
-    teamId: string;
-    teamName: string;
-    country: string;
-    firstAppareance: number;
-    constructorsChampionships: number;
-    driverChampionships: number;
-    url: string;    
-}
-
+import { Driver } from "../types/driver";
+import { Team } from "../types/team";
 
 export interface Circuit {
     circuitId: string;
@@ -98,24 +32,4 @@ export interface Race {
     circuit: Circuit;
     winner: Driver;
     teamWinner: Team;
-}
-
-export interface OpenF1Driver {
-    broadcast_name: string;
-    country_code: string;
-    driver_number: number;
-    first_name: string;
-    full_name: string;
-    headshot_url: string;
-    last_name: string;
-    meeting_key: number;
-    name_acronym: string;
-    session_key: number;
-    team_colour: string;
-    team_name: string;
-}
-
-export interface OpenF1Picture {
-    headshot_url: string;
-    last_name: string;
 }
