+++ lib/types/team.ts
@@ -1,0 +1,9 @@
+export interface Team {
+    teamId: string;
+    teamName: string;
+    teamNationality: string;
+    firstAppeareance: number;
+    constructorsChampionships: number;
+    driverChampionships: number;
+    url: string;
+}
